requirements.txt
asgiref==3.8.1
Django==5.1.3
psycopg2==2.9.10
sqlparse==0.5.2
tzdata==2024.2
djangorestframework==3.15.2
python-dotenv==1.0.1
django-cors-headers==4.6.0
drf-yasg==1.21.8
pillow==11.0.0
djangorestframework-simplejwt==5.3.1
django-filter==24.3
